+++ lib/src/lunar_calendar/models/models.dart
@@ -1,3 +1,4 @@
 export 'lunar_month.dart';
 export 'lunar_year.dart';
 export 'lunar_date.dart';
+export 'solar_term.dart';