+++ lib/src/utils/calculations.dart
@@ -1,5 +1,6 @@
 import 'package:chinese_lunar_calendar/chinese_lunar_calendar.dart';
 import 'package:chinese_lunar_calendar/src/constants/configs.dart';
+import 'package:chinese_lunar_calendar/src/constants/solar_terms.dart';
 
 // 计算汉字星期
 String getWeekDayCN(DateTime date) => weekDayCN[date.weekday - 1];
@@ -58,6 +59,11 @@ LunarDate getLunarDate(DateTime date) {
     spanDays = lunarYear.days + spanDays;
   }
   return lunarYear.getXthDay(spanDays);
+}
+
+// 获取当年节气
+List<SolarTerm> getSolarTerms(int year) {
+  return solarTermsTimeByYear['$year'] ?? List.empty();
 }
 
 // 计算六十天干地支